completed styling
