+++ public/index.html
@@ -24,7 +24,16 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+
+    <!-- Favicon and Title -->
+    <link rel="icon" type="image/svg+xml" sizes="any"
+    href="data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 100 100%22><text y=%22.9em%22 font-size=%2290%22>📃</text></svg>" />
+    <title>Public List</title>
+
+    <!-- Font Awesome Icons -->
+    <script src="https://kit.fontawesome.com/9de165e445.js" crossorigin="anonymous"></script>
+
+
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
